Move updating functions out of connector.go.
Because connector.go is getting *very* cluttered.
